test: split test suite into per resource files
Extract each Test* function from main_test.go into different files.

Also fix runTestCases to use require.NoError after json.Unmarshal, so
a malformed response fails the test cleanly instead of panicking when
validateFunc tries to read a nil map.

Fix #126.
